% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LM}
\alias{LM}
\title{Lindsey's Method}
\usage{
LM(df, n = 20, a = min(df), b = max(df), iter = 5000, burnin = 1000)
}
\arguments{
\item{df}{the sample to be estimated}

\item{n}{the number of equal width bins}

\item{a}{the minimum of the range to estimate the density}

\item{b}{the maximum of the range to estimate the density}

\item{iter}{the number of Gibbs samples}

\item{burnin}{the number of between-sample random numbers}
}
\value{
a list where \code{t} represents the possilble values and \code{lm} represents the density of that value
}
\description{
A density estimation using Lindsey's method with Gibbs sampling
}
\examples{
\dontrun{
rnR <- gibbsR(100,10)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
