% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PGM}
\alias{PGM}
\title{Penalized Gaussian Mixtures}
\usage{
PGM(x, K = 35, a0 = min(x), b0 = max(x), iter = 5000, burnin = 1000)
}
\arguments{
\item{x}{the sample to be estimated}

\item{K}{the number of Gaussian distributions to be cumulated}

\item{a0}{the minimum of the range to estimate the density}

\item{b0}{the maximum of the range to estimate the density}

\item{iter}{the number of Gibbs samples}

\item{burnin}{the number of between-sample random numbers}
}
\value{
a list where \code{t} represents the possilble values and \code{pgm} represents the density of that value
}
\description{
A density estimation using Penalized Gaussian Mixtures method with Gibbs sampling and Hamiltonian Monte Carlo sampling
}
\examples{
\dontrun{
rnR <- gibbsR(100,10)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
