\name{StatComp21080-package}
\alias{StatComp21080-package}
\alias{StatComp21080}
\docType{package}
\title{
  Density estimations using Lindsey's method and penalized Gaussian mixtures
}
\description{
   Density estimations using Lindsey's method and penalized Gaussian mixtures, where gibbs sampling and Hamiltonian Monte Carlo sampling.
}
\details{
  The function LM is to estimate the density of samples by Lindsey's method combined with Gibbs sampling.
  The function PGM is to estimate the density of samples by penalized Gaussian mixtures combined with Gibbs sampling and Hamiltonian Monte Carlo sampling.
}
\author{
Sun Jinglan
Maintainer: Sun Jinglan <sunjinglan2000@mail.ustc.edu.cn>
}

\keyword{ package }
